refactor(api): separate HTTP handlers by domain
