+++ apps/api/internal/httpapi/server.go
@@ -1,19 +1,10 @@
 package httpapi
 
 import (
-	"bytes"
 	"context"
-	"database/sql"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"io/fs"
-	"log"
 	"net/http"
-	"net/url"
-	"path"
-	"strconv"
 	"strings"
 	"time"
 
@@ -24,7 +15,6 @@ import (
 	"github.com/openclaw/clickclack/apps/api/internal/realtime"
 	"github.com/openclaw/clickclack/apps/api/internal/store"
 	"github.com/openclaw/clickclack/apps/api/internal/uploadstore"
-	"github.com/openclaw/clickclack/apps/api/internal/webassets"
 )
 
 type Server struct {
@@ -93,20 +83,6 @@ const (
 	passwordChangeLimit  = 5
 	passwordChangeWindow = 15 * time.Minute
 )
-
-var errAmbiguousCookie = errors.New("multiple cookies with the same name are not allowed")
-
-type actor struct {
-	user store.User
-	// sessionToken is the session this caller authenticated with, empty for
-	// every other way of resolving an actor: bot tokens, a trusted-proxy
-	// assertion, and the local development fallbacks. Handlers that revoke or
-	// revalidate the caller's own session key on it.
-	sessionToken string
-	botTokenID   string
-	workspaceID  string
-	scopes       []string
-}
 
 type Options struct {
 	UploadDir           string
@@ -299,1799 +275,6 @@ func (s *Server) Handler() http.Handler {
 	r.Head("/*", s.serveSPA)
 	r.Get("/*", s.serveSPA)
 	return r
-}
-
-func (s *Server) requireCookieCSRF(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if isSafeMethod(r.Method) || hasBearerAuth(r) || (!s.hasSessionCookie(r) && !s.hasAccessAssertion(r)) {
-			next.ServeHTTP(w, r)
-			return
-		}
-		if r.Header.Get(csrfHeaderName) != "1" || !s.sameOriginBrowserRequest(r) {
-			writeError(w, http.StatusForbidden, errors.New("cross-site session requests are not allowed"))
-			return
-		}
-		next.ServeHTTP(w, r)
-	})
-}
-
-func (s *Server) hasAccessAssertion(r *http.Request) bool {
-	return s.access != nil && r.Header.Get(accessAssertionHeader) != ""
-}
-
-func isSafeMethod(method string) bool {
-	switch method {
-	case http.MethodGet, http.MethodHead, http.MethodOptions, http.MethodTrace:
-		return true
-	default:
-		return false
-	}
-}
-
-func hasBearerAuth(r *http.Request) bool {
-	return strings.HasPrefix(r.Header.Get("Authorization"), "Bearer ")
-}
-
-func (s *Server) hasSessionCookie(r *http.Request) bool {
-	cookies := r.CookiesNamed(s.cookies.Session)
-	if len(cookies) > 1 {
-		return true
-	}
-	return len(cookies) == 1 && cookies[0].Value != ""
-}
-
-func requestCookie(r *http.Request, name string) (*http.Cookie, error) {
-	cookies := r.CookiesNamed(name)
-	switch len(cookies) {
-	case 0:
-		return nil, http.ErrNoCookie
-	case 1:
-		return cookies[0], nil
-	default:
-		return nil, errAmbiguousCookie
-	}
-}
-
-// AccessLogMode selects how much of the per-request access log the server
-// writes. The zero value logs every request, which is what operators got
-// before the mode existed.
-type AccessLogMode string
-
-const (
-	AccessLogAll    AccessLogMode = "all"
-	AccessLogErrors AccessLogMode = "errors"
-	AccessLogOff    AccessLogMode = "off"
-)
-
-type pathOnlyLogFormatter struct {
-	Logger middleware.LoggerInterface
-	Mode   AccessLogMode
-}
-
-func (f *pathOnlyLogFormatter) NewLogEntry(r *http.Request) middleware.LogEntry {
-	mode := f.mode()
-	if mode == AccessLogOff {
-		return &pathOnlyLogEntry{logger: f.logger(), request: r, mode: mode}
-	}
-	scheme := "http"
-	if r.TLS != nil {
-		scheme = "https"
-	}
-	prefix := fmt.Sprintf("method=%q scheme=%q host=%q proto=%q remote=%q correlation_id=%q ", r.Method, scheme, r.Host, r.Proto, r.RemoteAddr, correlationIDFromContext(r.Context()))
-	return &pathOnlyLogEntry{logger: f.logger(), prefix: prefix, request: r, mode: mode}
-}
-
-func (f *pathOnlyLogFormatter) logger() middleware.LoggerInterface {
-	if f.Logger != nil {
-		return f.Logger
-	}
-	return log.Default()
-}
-
-func (f *pathOnlyLogFormatter) mode() AccessLogMode {
-	if f.Mode == "" {
-		return AccessLogAll
-	}
-	return f.Mode
-}
-
-type pathOnlyLogEntry struct {
-	logger  middleware.LoggerInterface
-	prefix  string
-	request *http.Request
-	mode    AccessLogMode
-}
-
-func (e *pathOnlyLogEntry) Write(status, bytes int, _ http.Header, elapsed time.Duration, _ interface{}) {
-	switch e.mode {
-	case AccessLogOff:
-		return
-	case AccessLogErrors:
-		if status < http.StatusBadRequest {
-			return
-		}
-	}
-	route := chi.RouteContext(e.request.Context()).RoutePattern()
-	if route == "" {
-		route = "unmatched"
-	}
-	e.logger.Print(fmt.Sprintf("%sroute=%q status=%03d bytes=%d elapsed=%s", e.prefix, route, status, bytes, elapsed))
-}
-
-func (e *pathOnlyLogEntry) Panic(v interface{}, _ []byte) {
-	middleware.PrintPrettyStack(v)
-}
-
-func (s *Server) resolveRoute(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	workspaceRouteID := chi.URLParam(r, "workspace_route_id")
-	targetRouteID := chi.URLParam(r, "target_route_id")
-	scope := routeScopeForParam(targetRouteID)
-	if scope == "" {
-		writeError(w, http.StatusNotFound, errors.New("route not found"))
-		return
-	}
-	if err := act.requireScope(scope); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var target store.RouteTarget
-	if isLegacyRouteParam(workspaceRouteID) || isLegacyRouteParam(targetRouteID) {
-		target, err = s.store.ResolveLegacyRouteTarget(r.Context(), act.user.ID, workspaceRouteID, targetRouteID)
-	} else {
-		target, err = s.store.ResolveRouteTarget(r.Context(), act.user.ID, workspaceRouteID, targetRouteID)
-	}
-	if err != nil {
-		writeError(w, http.StatusNotFound, errors.New("route not found"))
-		return
-	}
-	if err := act.requireWorkspace(target.WorkspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if scope != routeScopeForTargetType(target.TargetType) {
-		writeError(w, http.StatusNotFound, errors.New("route not found"))
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"route": target})
-}
-
-func routeScopeForParam(value string) string {
-	switch {
-	case strings.HasPrefix(value, "C"), strings.HasPrefix(value, "chn_"):
-		return "channels:read"
-	case strings.HasPrefix(value, "D"), strings.HasPrefix(value, "dm_"):
-		return "dms:read"
-	case strings.HasPrefix(value, "M"), strings.HasPrefix(value, "msg_"):
-		return "threads:read"
-	default:
-		return ""
-	}
-}
-
-func routeScopeForTargetType(targetType string) string {
-	switch targetType {
-	case "channel":
-		return "channels:read"
-	case "direct":
-		return "dms:read"
-	case "thread":
-		return "threads:read"
-	default:
-		return ""
-	}
-}
-
-func isLegacyRouteParam(value string) bool {
-	return strings.HasPrefix(value, "wsp_") ||
-		strings.HasPrefix(value, "chn_") ||
-		strings.HasPrefix(value, "dm_") ||
-		strings.HasPrefix(value, "msg_")
-}
-
-func (s *Server) me(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("profile:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	preferences, err := s.store.GetAppearancePreferences(r.Context(), act.user.ID)
-	payload := currentUserPayload{User: act.user, AppearancePreferences: preferences}
-	if err == nil {
-		payload.PasswordEnrolled, err = s.passwordEnrolled(r.Context(), act.user.ID)
-	}
-	writeResult(w, map[string]any{"user": payload}, err)
-}
-
-func (s *Server) updateMe(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if act.botTokenID != "" {
-		writeError(w, http.StatusForbidden, errors.New("bot tokens cannot update profiles"))
-		return
-	}
-	var body struct {
-		DisplayName           *string                           `json:"display_name"`
-		Handle                *string                           `json:"handle"`
-		AvatarURL             *string                           `json:"avatar_url"`
-		NotificationSettings  *store.NotificationSettings       `json:"notification_settings"`
-		AppearancePreferences *store.AppearancePreferencesPatch `json:"appearance_preferences"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	updated, err := s.store.UpdateCurrentUser(r.Context(), store.UpdateCurrentUserInput{
-		UserID:                act.user.ID,
-		DisplayName:           body.DisplayName,
-		Handle:                body.Handle,
-		AvatarURL:             body.AvatarURL,
-		NotificationSettings:  body.NotificationSettings,
-		AppearancePreferences: body.AppearancePreferences,
-	})
-	payload := currentUserPayload{User: updated.User, AppearancePreferences: updated.AppearancePreferences}
-	if err == nil {
-		payload.PasswordEnrolled, err = s.passwordEnrolled(r.Context(), updated.User.ID)
-	}
-	writeResult(w, map[string]any{"user": payload}, err)
-}
-
-type currentUserPayload struct {
-	store.User
-	AppearancePreferences *store.AppearancePreferences `json:"appearance_preferences,omitempty"`
-	PasswordEnrolled      bool                         `json:"password_enrolled"`
-}
-
-// passwordEnrolled reports whether an account has a password on file. The SPA
-// pairs it with the advertised auth methods to decide whether to offer the
-// change-password form. It is reported only for the caller's own account, so it
-// discloses nothing about who else can sign in with a password.
-func (s *Server) passwordEnrolled(ctx context.Context, userID string) (bool, error) {
-	hash, err := s.store.GetUserPasswordHash(ctx, userID)
-	return hash != "", err
-}
-
-func (s *Server) listWorkspaces(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("workspaces:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	items, err := s.store.ListWorkspaces(r.Context(), act.user.ID)
-	if err == nil && act.botTokenID != "" {
-		filtered := items[:0]
-		for _, item := range items {
-			if item.ID == act.workspaceID {
-				filtered = append(filtered, item)
-			}
-		}
-		items = filtered
-	}
-	writeResult(w, map[string]any{"workspaces": items}, err)
-}
-
-func (s *Server) createWorkspace(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if act.botTokenID != "" {
-		writeError(w, http.StatusForbidden, errors.New("bot tokens cannot create workspaces"))
-		return
-	}
-	var body struct {
-		Name string `json:"name"`
-		Slug string `json:"slug"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	workspaces, err := s.store.ListWorkspaces(r.Context(), act.user.ID)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	hasNonGuestMembership, err := s.store.UserHasNonGuestMembership(r.Context(), act.user.ID)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	if len(workspaces) > 0 && !hasNonGuestMembership {
-		writeError(w, http.StatusForbidden, store.ErrModerationRestricted)
-		return
-	}
-	workspace, err := s.store.CreateWorkspace(r.Context(), store.CreateWorkspaceInput{Name: body.Name, Slug: body.Slug}, act.user.ID)
-	writeResultStatus(w, http.StatusCreated, map[string]any{"workspace": workspace}, err)
-}
-
-func (s *Server) getWorkspace(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireScope("workspaces:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	workspace, err := s.store.GetWorkspace(r.Context(), workspaceID, act.user.ID)
-	writeResult(w, map[string]any{"workspace": workspace}, err)
-}
-
-func (s *Server) updateWorkspace(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if act.botTokenID != "" {
-		writeError(w, http.StatusForbidden, errors.New("bot tokens cannot update workspaces"))
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireScope("workspaces:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var body struct {
-		Name    *string `json:"name"`
-		Slug    *string `json:"slug"`
-		IconURL *string `json:"icon_url"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	if body.Name == nil && body.Slug == nil && body.IconURL == nil {
-		writeError(w, http.StatusBadRequest, errors.New("workspace update requires at least one field"))
-		return
-	}
-	workspace, event, err := s.store.UpdateWorkspace(r.Context(), store.UpdateWorkspaceInput{
-		WorkspaceID: workspaceID,
-		ActorUserID: act.user.ID,
-		Name:        body.Name,
-		Slug:        body.Slug,
-		IconURL:     body.IconURL,
-	})
-	if err == nil && event.ID != "" {
-		s.publishEvent(r.Context(), event)
-		s.recordAudit(r.Context(), workspaceID, act.user.ID, "workspace.updated", "workspace", workspaceID, map[string]any{"name": workspace.Name, "slug": workspace.Slug, "icon_url": workspace.IconURL})
-	}
-	writeResult(w, map[string]any{"workspace": workspace, "event": event}, err)
-}
-
-func (s *Server) transferWorkspaceOwnership(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if act.botTokenID != "" {
-		writeError(w, http.StatusForbidden, errors.New("bot tokens cannot transfer workspace ownership"))
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireScope("workspaces:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var body struct {
-		UserID string `json:"user_id"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	workspace, event, err := s.store.TransferWorkspaceOwnership(r.Context(), store.TransferWorkspaceOwnershipInput{
-		WorkspaceID:    workspaceID,
-		ActorUserID:    act.user.ID,
-		NewOwnerUserID: body.UserID,
-	})
-	if err == nil && event.ID != "" {
-		s.publishEvent(r.Context(), event)
-		s.recordAudit(r.Context(), workspaceID, act.user.ID, "workspace.ownership_transferred", "workspace", workspaceID, map[string]any{"new_owner_user_id": body.UserID})
-	}
-	writeResult(w, map[string]any{"workspace": workspace, "event": event}, err)
-}
-
-func (s *Server) deleteWorkspace(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if act.botTokenID != "" {
-		writeError(w, http.StatusForbidden, errors.New("bot tokens cannot delete workspaces"))
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireScope("workspaces:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	cleanups, err := s.store.DeleteWorkspace(r.Context(), workspaceID, act.user.ID)
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	if err := s.cleanupUploadObjects(r.Context(), cleanups); err != nil {
-		log.Printf("workspace %s deleted with pending upload cleanup retry: %v", workspaceID, err)
-	}
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func (s *Server) CleanupPendingUploadObjects(ctx context.Context, limit int) error {
-	if limit <= 0 {
-		limit = uploadCleanupSweepLimit
-	}
-	attempted := make(map[string]struct{})
-	var cleanupErrors []error
-	queryLimit := limit
-	for {
-		cleanups, err := s.store.ListPendingUploadCleanups(ctx, queryLimit)
-		if err != nil {
-			return err
-		}
-		if len(cleanups) == 0 {
-			return errors.Join(cleanupErrors...)
-		}
-		pending := cleanups[:0]
-		for _, cleanup := range cleanups {
-			if _, seen := attempted[cleanup.ID]; seen {
-				continue
-			}
-			attempted[cleanup.ID] = struct{}{}
-			pending = append(pending, cleanup)
-		}
-		if len(pending) == 0 {
-			if len(cleanups) < queryLimit {
-				return errors.Join(cleanupErrors...)
-			}
-			queryLimit += limit
-			continue
-		}
-		if err := s.cleanupUploadObjects(ctx, pending); err != nil {
-			cleanupErrors = append(cleanupErrors, err)
-		}
-	}
-}
-
-func (s *Server) cleanupUploadObjects(ctx context.Context, cleanups []store.PendingUploadCleanup) error {
-	if len(cleanups) == 0 {
-		return nil
-	}
-	if s.uploadStorage == nil {
-		err := errors.New("upload storage is not configured")
-		for _, cleanup := range cleanups {
-			_ = s.store.RecordPendingUploadCleanupFailure(ctx, cleanup.ID, err.Error())
-		}
-		return err
-	}
-	var cleanupErrors []error
-	for _, cleanup := range cleanups {
-		if err := s.uploadStorage.Delete(ctx, cleanup.StoragePath); err != nil {
-			_ = s.store.RecordPendingUploadCleanupFailure(ctx, cleanup.ID, err.Error())
-			cleanupErrors = append(cleanupErrors, err)
-			continue
-		}
-		if err := s.store.DeletePendingUploadCleanup(ctx, cleanup.ID); err != nil {
-			cleanupErrors = append(cleanupErrors, err)
-		}
-	}
-	return errors.Join(cleanupErrors...)
-}
-
-func (s *Server) listWorkspaceMemberPage(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireScope("workspaces:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	page, err := parseWorkspaceMemberPageRequest(r)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	members, err := s.store.ListWorkspaceMemberPage(r.Context(), workspaceID, act.user.ID, page)
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, members)
-}
-
-func parseWorkspaceMemberPageRequest(r *http.Request) (store.WorkspaceMemberPageRequest, error) {
-	values := r.URL.Query()
-	page := store.WorkspaceMemberPageRequest{
-		Cursor: values.Get("cursor"),
-		Query:  values.Get("q"),
-		Role:   values.Get("role"),
-	}
-	if rawLimit := strings.TrimSpace(values.Get("limit")); rawLimit != "" {
-		limit, err := strconv.ParseInt(rawLimit, 10, 32)
-		if err != nil || limit < 1 {
-			return page, fmt.Errorf("%w: limit must be positive", store.ErrInvalidWorkspaceMemberPage)
-		}
-		page.Limit = int(limit)
-	}
-	return page, nil
-}
-
-func (s *Server) listWorkspaceMembers(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireScope("workspaces:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	members, err := s.store.ListWorkspaceMembers(r.Context(), workspaceID, act.user.ID)
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"members": members})
-}
-
-func (s *Server) updateWorkspaceMemberModeration(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireScope("workspaces:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var body struct {
-		Role           string  `json:"role"`
-		TimeoutUntil   string  `json:"timeout_until"`
-		TimeoutMinutes int     `json:"timeout_minutes"`
-		ClearTimeout   bool    `json:"clear_timeout"`
-		Blocked        *bool   `json:"blocked"`
-		ModerationNote *string `json:"moderation_note"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	timeoutUntil := optionalString(body.TimeoutUntil)
-	if timeoutUntil == nil && body.TimeoutMinutes > 0 {
-		value := time.Now().Add(time.Duration(body.TimeoutMinutes) * time.Minute).UTC().Format(time.RFC3339Nano)
-		timeoutUntil = &value
-	}
-	member, event, err := s.store.UpdateMemberModeration(r.Context(), store.UpdateMemberModerationInput{
-		WorkspaceID:    workspaceID,
-		TargetUserID:   chi.URLParam(r, "user_id"),
-		ActorUserID:    act.user.ID,
-		Role:           body.Role,
-		TimeoutUntil:   timeoutUntil,
-		ClearTimeout:   body.ClearTimeout,
-		Blocked:        body.Blocked,
-		ModerationNote: body.ModerationNote,
-	})
-	if err == nil && event.ID != "" {
-		s.hub.Publish(event)
-	}
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"member": member, "event": event})
-}
-
-func (s *Server) listChannels(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireScope("channels:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	channels, err := s.store.ListChannels(r.Context(), workspaceID, act.user.ID)
-	writeResult(w, map[string]any{"channels": channels}, err)
-}
-
-func (s *Server) createChannel(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("channels:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(chi.URLParam(r, "workspace_id")); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var body struct {
-		Name            string `json:"name"`
-		DisplayTitle    string `json:"display_title"`
-		Kind            string `json:"kind"`
-		ExternalManaged bool   `json:"external_managed"`
-		ExternalRef     string `json:"external_ref"`
-		ExternalURL     string `json:"external_url"`
-		SidebarSection  string `json:"sidebar_section"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	channel, event, err := s.store.CreateChannel(r.Context(), store.CreateChannelInput{
-		WorkspaceID:     chi.URLParam(r, "workspace_id"),
-		Name:            body.Name,
-		DisplayTitle:    body.DisplayTitle,
-		Kind:            body.Kind,
-		UserID:          act.user.ID,
-		ExternalManaged: body.ExternalManaged,
-		ExternalRef:     body.ExternalRef,
-		ExternalURL:     body.ExternalURL,
-		SidebarSection:  body.SidebarSection,
-	})
-	if err == nil {
-		s.publishEvent(r.Context(), event)
-	}
-	writeResultStatus(w, http.StatusCreated, map[string]any{"channel": channel, "event": event}, err)
-}
-
-func (s *Server) listTopics(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("channels:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	topics, err := s.store.ListTopics(r.Context(), workspaceID, act.user.ID)
-	writeResult(w, map[string]any{"topics": topics}, err)
-}
-
-func (s *Server) createTopic(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("channels:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	workspaceID := chi.URLParam(r, "workspace_id")
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var body struct {
-		ChannelID string `json:"channel_id"`
-		Name      string `json:"name"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	topic, err := s.store.CreateTopic(r.Context(), store.CreateTopicInput{WorkspaceID: workspaceID, ChannelID: body.ChannelID, Name: body.Name, CreatedBy: act.user.ID})
-	writeResultStatus(w, http.StatusCreated, map[string]any{"topic": topic}, err)
-}
-
-func (s *Server) listMessages(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("messages:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	page, err := parseMessagePageRequest(r)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	if values, ok := r.URL.Query()["topic_id"]; ok {
-		if len(values) == 0 || strings.TrimSpace(values[0]) == "" {
-			writeError(w, http.StatusBadRequest, fmt.Errorf("%w: topic_id is required", store.ErrInvalidMessagePage))
-			return
-		}
-		page.TopicID = strings.TrimSpace(values[0])
-	}
-	if !s.requireBotChannelWorkspace(w, r, act, chi.URLParam(r, "channel_id")) {
-		return
-	}
-	messages, err := s.store.ListMessages(r.Context(), chi.URLParam(r, "channel_id"), act.user.ID, page)
-	writeMessagePage(w, messages, err)
-}
-
-func (s *Server) createMessage(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("messages:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var body struct {
-		Body            string `json:"body"`
-		QuotedMessageID string `json:"quoted_message_id"`
-		Nonce           string `json:"nonce"`
-		TopicID         string `json:"topic_id"`
-		UploadID        string `json:"upload_id"`
-		Kind            string `json:"kind"`
-		TurnID          string `json:"turn_id"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	kind, turnID, ok := s.resolveMessageKind(w, act, body.Kind, body.TurnID)
-	if !ok {
-		return
-	}
-	if !s.requireBotChannelWorkspace(w, r, act, chi.URLParam(r, "channel_id")) {
-		return
-	}
-	if !s.requireCreateUpload(w, r, act, body.UploadID, body.Nonce, chi.URLParam(r, "channel_id"), "") {
-		return
-	}
-	message, event, err := s.store.CreateMessage(r.Context(), store.CreateMessageInput{ChannelID: chi.URLParam(r, "channel_id"), AuthorID: act.user.ID, Body: body.Body, QuotedMessageID: optionalString(body.QuotedMessageID), Nonce: body.Nonce, TopicID: body.TopicID, UploadID: body.UploadID, Kind: kind, TurnID: turnID})
-	if err == nil && event.ID != "" {
-		s.publishEvent(r.Context(), event)
-		if !store.IsActivityMessageKind(message.Kind) {
-			s.notifyMessageCreated(r.Context(), message, event.MentionedUserIDs)
-		}
-	}
-	writeMessageCreateResult(w, message, event, err)
-}
-
-// resolveMessageKind validates a caller-supplied message kind and turn_id and
-// enforces the activity authorization contract:
-//
-//   - an empty/'message' kind is always allowed and returned as 'message',
-//   - an unknown kind is a 400,
-//   - an ordinary 'message' MUST NOT carry a turn_id (400); turn_id correlates
-//     agent activity rows only, so a non-empty value on an ordinary message is
-//     a client contract violation and fails closed,
-//   - an activity kind (agent_commentary/agent_tool) requires a BOT token that
-//     carries agent_activity:write; a human session always gets 403 and a bot
-//     without the scope gets 403, and may carry a turn_id.
-//
-// It writes the error response itself and returns ok=false when the request
-// must not proceed. On success it returns the normalized kind and the turn_id
-// that should be persisted.
-func (s *Server) resolveMessageKind(w http.ResponseWriter, act actor, rawKind, rawTurnID string) (string, string, bool) {
-	kind, err := store.NormalizeMessageKind(rawKind)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return "", "", false
-	}
-	if !store.IsActivityMessageKind(kind) {
-		if rawTurnID != "" {
-			writeError(w, http.StatusBadRequest, store.ErrTurnIDNotAllowed)
-			return "", "", false
-		}
-		return kind, "", true
-	}
-	if act.botTokenID == "" {
-		writeError(w, http.StatusForbidden, errors.New("agent activity messages require a bot token"))
-		return "", "", false
-	}
-	if err := act.requireScope(store.AgentActivityWriteScope); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return "", "", false
-	}
-	return kind, rawTurnID, true
-}
-
-func (s *Server) getMessage(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("messages:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	message, ok := s.requireBotMessageResource(w, r, act, chi.URLParam(r, "message_id"), "dms:read")
-	if !ok {
-		return
-	}
-	if act.botTokenID == "" {
-		message, err = s.store.GetMessage(r.Context(), chi.URLParam(r, "message_id"), act.user.ID)
-	}
-	writeResult(w, map[string]any{"message": message}, err)
-}
-
-func (s *Server) getMessageByNonce(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("X-ClickClack-Message-Nonce", "supported")
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("messages:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	workspaceID := strings.TrimSpace(r.URL.Query().Get("workspace_id"))
-	if workspaceID == "" {
-		writeError(w, http.StatusBadRequest, errors.New("workspace_id is required"))
-		return
-	}
-	nonce, err := normalizeClientNonce(r.URL.Query().Get("nonce"))
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	if nonce == "" {
-		writeError(w, http.StatusBadRequest, errors.New("nonce is required"))
-		return
-	}
-	if !s.authorizeWorkspaceAccess(w, r, act, workspaceID) {
-		return
-	}
-	message, err := s.store.GetMessageByNonce(r.Context(), act.user.ID, nonce)
-	switch {
-	case err == nil && message.WorkspaceID != workspaceID:
-		writeStoreError(w, store.ErrClientNonceConflict)
-	case err == nil:
-		if !requireBotMessageDirectScope(w, act, message, "dms:write") {
-			return
-		}
-		writeJSON(w, http.StatusOK, map[string]any{"message": message})
-	case errors.Is(err, sql.ErrNoRows):
-		writeError(w, http.StatusNotFound, err)
-	default:
-		writeStoreError(w, err)
-	}
-}
-
-func (s *Server) ensureMessageRoute(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("threads:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	messageID := chi.URLParam(r, "message_id")
-	if _, ok := s.requireBotMessageResource(w, r, act, messageID, "dms:read"); !ok {
-		return
-	}
-	message, err := s.store.EnsureMessageRouteID(r.Context(), act.user.ID, messageID)
-	if errors.Is(err, sql.ErrNoRows) || errors.Is(err, store.ErrModerationRestricted) {
-		writeError(w, http.StatusNotFound, errors.New("message not found"))
-		return
-	}
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"message": message})
-}
-
-func (s *Server) getThread(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("threads:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if _, ok := s.requireBotMessageResource(w, r, act, chi.URLParam(r, "message_id"), "dms:read"); !ok {
-		return
-	}
-	if r.URL.Query().Has("mode") {
-		writeError(w, http.StatusBadRequest, errors.New("use latest or a thread sequence cursor"))
-		return
-	}
-	req, err := parseMessagePageRequest(r)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	latest := strings.TrimSpace(r.URL.Query().Get("latest"))
-	if latest != "" && latest != "true" && latest != "false" {
-		writeError(w, http.StatusBadRequest, errors.New("latest must be true or false"))
-		return
-	}
-	page, err := s.store.GetThreadPage(r.Context(), chi.URLParam(r, "message_id"), act.user.ID, store.ThreadPageRequest{MessagePageRequest: req, Latest: latest == "true"})
-	writeResult(w, page, err)
-}
-
-func (s *Server) createThreadReply(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("threads:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var body struct {
-		Body            string `json:"body"`
-		QuotedMessageID string `json:"quoted_message_id"`
-		Nonce           string `json:"nonce"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	if _, ok := s.requireBotMessageResource(w, r, act, chi.URLParam(r, "message_id"), "dms:write"); !ok {
-		return
-	}
-	message, state, events, err := s.store.CreateThreadReply(r.Context(), store.CreateThreadReplyInput{RootMessageID: chi.URLParam(r, "message_id"), AuthorID: act.user.ID, Body: body.Body, QuotedMessageID: optionalString(body.QuotedMessageID), Nonce: body.Nonce})
-	if err == nil && len(events) > 0 {
-		s.publishEvents(r.Context(), events)
-		s.notifyMessageCreated(r.Context(), message, messageEventMentionedUserIDs(events))
-	}
-	writeThreadReplyCreateResult(w, message, state, events, err)
-}
-
-func (s *Server) addReaction(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("messages:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	var body struct {
-		Emoji string `json:"emoji"`
-	}
-	if err := readJSON(w, r, &body); err != nil {
-		writeError(w, http.StatusBadRequest, err)
-		return
-	}
-	if _, ok := s.requireBotMessageResource(w, r, act, chi.URLParam(r, "message_id"), "dms:write"); !ok {
-		return
-	}
-	event, err := s.store.AddReaction(r.Context(), store.CreateReactionInput{MessageID: chi.URLParam(r, "message_id"), UserID: act.user.ID, Emoji: body.Emoji})
-	if err == nil && event.ID != "" {
-		s.publishEvent(r.Context(), event)
-	}
-	s.writeReactionMutationResult(w, r, http.StatusCreated, act.user.ID, chi.URLParam(r, "message_id"), event, err)
-}
-
-func (s *Server) removeReaction(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("messages:write"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if _, ok := s.requireBotMessageResource(w, r, act, chi.URLParam(r, "message_id"), "dms:write"); !ok {
-		return
-	}
-	emoji := chi.URLParam(r, "emoji")
-	// Chi routes on RawPath when present; otherwise the parameter is already decoded.
-	if r.URL.RawPath != "" {
-		emoji, err = url.PathUnescape(emoji)
-		if err != nil {
-			writeError(w, http.StatusBadRequest, err)
-			return
-		}
-	}
-	event, err := s.store.RemoveReaction(r.Context(), store.CreateReactionInput{MessageID: chi.URLParam(r, "message_id"), UserID: act.user.ID, Emoji: emoji})
-	if err == nil && event.ID != "" {
-		s.publishEvent(r.Context(), event)
-	}
-	s.writeReactionMutationResult(w, r, http.StatusOK, act.user.ID, chi.URLParam(r, "message_id"), event, err)
-}
-
-func (s *Server) listEvents(w http.ResponseWriter, r *http.Request) {
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	workspaceID := r.URL.Query().Get("workspace_id")
-	if err := act.requireScope("realtime:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	result := map[string]any{}
-	if r.URL.Query().Get("include_tail") == "true" {
-		tailCursor, err := s.store.LatestEventCursor(r.Context(), workspaceID, act.user.ID)
-		if err != nil {
-			writeResult(w, nil, err)
-			return
-		}
-		result["tail_cursor"] = tailCursor
-	}
-	events, err := s.store.ListEventsAfter(r.Context(), workspaceID, act.user.ID, r.URL.Query().Get("after_cursor"), queryInt(r, "limit", 200))
-	if err == nil {
-		events = filterEventsForUser(events, act.user.ID)
-		result["events"] = events
-	}
-	writeResult(w, result, err)
-}
-
-func (s *Server) websocket(w http.ResponseWriter, r *http.Request) {
-	bearerProtocol := websocketBearerProtocol(r)
-	if r.Header.Get("Authorization") == "" {
-		if bearerProtocol != "" {
-			r.Header.Set("Authorization", "Bearer "+strings.TrimPrefix(bearerProtocol, websocketBearerProtocolPrefix))
-		}
-	}
-	act, err := s.currentActor(r)
-	if err != nil {
-		writeError(w, http.StatusUnauthorized, err)
-		return
-	}
-	if err := act.requireScope("realtime:read"); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	workspaceID := r.URL.Query().Get("workspace_id")
-	if workspaceID == "" {
-		writeError(w, http.StatusBadRequest, errors.New("workspace_id is required"))
-		return
-	}
-	if err := act.requireWorkspace(workspaceID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	if _, err := s.store.GetWorkspace(r.Context(), workspaceID, act.user.ID); err != nil {
-		writeError(w, http.StatusForbidden, err)
-		return
-	}
-	subscription, unsubscribe := s.hub.Subscribe(workspaceID)
-	defer unsubscribe()
-	acceptOptions := &websocket.AcceptOptions{OriginPatterns: s.websocketOriginPatterns(r)}
-	if bearerProtocol != "" {
-		acceptOptions.Subprotocols = []string{bearerProtocol}
-	}
-	conn, err := websocket.Accept(w, r, acceptOptions)
-	if err != nil {
-		return
-	}
-	defer conn.CloseNow()
-	ctx := conn.CloseRead(r.Context())
-	replayCursor := r.URL.Query().Get("after_cursor")
-	// Revalidate the exact credential in the shared store: setup replay can replace
-	// a bot secret without changing its token ID, including on another replica.
-	bearerToken := strings.TrimSpace(strings.TrimPrefix(r.Header.Get("Authorization"), "Bearer "))
-	credentialAuthorityLive := func() bool {
-		var err error
-		revokedReason := realtimeSessionRevokedCloseReason
-		if act.botTokenID != "" {
-			_, err = s.store.GetBotTokenAuth(ctx, bearerToken)
-			revokedReason = "bot token revoked; reconnect with a valid token"
-		} else if act.sessionToken != "" {
-			_, err = s.sessionUser(ctx, act.sessionToken)
-		}
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) || errors.Is(err, store.ErrSessionExpired) {
-				_ = conn.Close(websocket.StatusPolicyViolation, revokedReason)
-			} else {
-				_ = conn.Close(websocket.StatusTryAgainLater, "credential verification unavailable; retry")
-			}
-			return false
-		}
-		return true
-	}
-	writeEvent := func(event store.Event) bool {
-		return credentialAuthorityLive() && writeWS(ctx, conn, event) == nil
-	}
-	sessionRecheck := time.NewTicker(s.realtimeSessionCheck)
-	defer sessionRecheck.Stop()
-	// Startup and live delivery share one ordered, authorized durable-log drain.
-	// Capture a finite tail each time; a wake received during this drain stays queued.
-	drain := func() bool {
-		if !credentialAuthorityLive() {
-			return false
-		}
-		replayTail, err := s.store.LatestEventCursor(ctx, workspaceID, act.user.ID)
-		if err != nil {
-			_ = conn.Close(websocket.StatusTryAgainLater, realtimeReplayCloseReason)
-			return false
-		}
-		if replayCursor != "" {
-			exists, err := s.store.EventCursorExists(ctx, workspaceID, act.user.ID, replayCursor)
-			if err != nil {
-				_ = conn.Close(websocket.StatusTryAgainLater, realtimeReplayCloseReason)
-				return false
-			}
-			if !exists {
-				_ = conn.Close(realtimeResyncRequiredStatus, realtimeResyncRequiredCloseReason)
-				return false
-			}
-		}
-		if replayCursor != "" && (replayTail == "" || replayCursor > replayTail) {
-			_ = conn.Close(realtimeResyncRequiredStatus, realtimeResyncRequiredCloseReason)
-			return false
-		}
-		replayedEvents := 0
-		for replayTail != "" && replayCursor < replayTail {
-			pageCursor := replayCursor
-			backlog, err := s.store.ListEventsAfter(ctx, workspaceID, act.user.ID, pageCursor, realtimeReplayPageSize)
-			if err != nil {
-				_ = conn.Close(websocket.StatusTryAgainLater, realtimeReplayCloseReason)
-				return false
-			}
-			if len(backlog) == 0 {
-				_ = conn.Close(realtimeResyncRequiredStatus, realtimeResyncRequiredCloseReason)
-				return false
-			}
-			if pageCursor != "" {
-				exists, err := s.store.EventCursorExists(ctx, workspaceID, act.user.ID, pageCursor)
-				if err != nil {
-					_ = conn.Close(websocket.StatusTryAgainLater, realtimeReplayCloseReason)
-					return false
-				}
-				if !exists {
-					_ = conn.Close(realtimeResyncRequiredStatus, realtimeResyncRequiredCloseReason)
-					return false
-				}
-			}
-			previousCursor := pageCursor
-			for _, event := range backlog {
-				if event.Cursor > replayTail {
-					break
-				}
-				if replayedEvents >= s.realtimeReplayLimit {
-					_ = conn.Close(realtimeResyncRequiredStatus, realtimeResyncRequiredCloseReason)
-					return false
-				}
-				replayedEvents++
-				// ListEventsAfter prefilters visibility, while this live lookup closes
-				// the revocation window between fetching a page and writing its events.
-				deliver, err := s.shouldDeliverEventToActorResult(ctx, event, act.user.ID)
-				if err != nil {
-					_ = conn.Close(websocket.StatusTryAgainLater, realtimeReplayCloseReason)
-					return false
-				}
-				if !deliver {
-					replayCursor = event.Cursor
-					continue
-				}
-				if !writeEvent(event) {
-					return false
-				}
-				replayCursor = event.Cursor
-			}
-			if replayCursor == previousCursor {
-				_ = conn.Close(websocket.StatusTryAgainLater, realtimeReplayCloseReason)
-				return false
-			}
-		}
-		return true
-	}
-	if !drain() {
-		return
-	}
-	for {
-		// Prefer overflow termination to starting another durable drain.
-		select {
-		case <-subscription.Done:
-			_ = conn.Close(websocket.StatusTryAgainLater, realtimeOverflowCloseReason)
-			return
-		default:
-		}
-		select {
-		case <-ctx.Done():
-			return
-		case <-subscription.Done:
-			_ = conn.Close(websocket.StatusTryAgainLater, realtimeOverflowCloseReason)
-			return
-		case <-sessionRecheck.C:
-			// An idle socket delivers nothing to revalidate against, so revocation
-			// reaches it here instead of waiting for the workspace's next event.
-			if !credentialAuthorityLive() {
-				return
-			}
-		case _, ok := <-subscription.Wake:
-			if !ok {
-				continue
-			}
-			if !drain() {
-				return
-			}
-		case event, ok := <-subscription.Events:
-			if !ok {
-				continue
-			}
-			// Revocations and other cursorless events are intentionally ephemeral.
-			if eventRevokesWorkspaceAccess(event, act.user.ID) {
-				_ = conn.Close(websocket.StatusPolicyViolation, "workspace access revoked")
-				return
-			}
-			if !s.shouldDeliverEventToActor(ctx, event, act.user.ID) {
-				continue
-			}
-			if !writeEvent(event) {
-				return
-			}
-		}
-	}
-}
-
-func websocketBearerProtocol(r *http.Request) string {
-	for _, protocol := range strings.Split(r.Header.Get("Sec-WebSocket-Protocol"), ",") {
-		protocol = strings.TrimSpace(protocol)
-		if strings.HasPrefix(protocol, websocketBearerProtocolPrefix) {
-			return protocol
-		}
-	}
-	return ""
-}
-
-func (s *Server) websocketOriginPatterns(r *http.Request) []string {
-	publicURL, err := url.Parse(strings.TrimSpace(firstNonEmpty(s.frontendURL, s.githubOAuth.PublicURL)))
-	if err != nil || publicURL.Host == "" {
-		return nil
-	}
-	return []string{publicURL.Scheme + "://" + publicURL.Host}
-}
-
-// shouldDeliverEvent gates per-user-private events so they only reach allowed
-// sessions and never leak to other workspace members.
-func shouldDeliverEvent(event store.Event, userID string) bool {
-	if len(event.RecipientUserIDs) > 0 {
-		for _, allowed := range event.RecipientUserIDs {
-			if allowed == userID {
-				return true
-			}
-		}
-		return false
-	}
-	switch event.Type {
-	case "channel.read", "dm.read":
-		payload, ok := event.Payload.(map[string]string)
-		if !ok {
-			// Backlog payloads come back via ListEventsAfter as map[string]any.
-			if anyPayload, ok := event.Payload.(map[string]any); ok {
-				if v, _ := anyPayload["user_id"].(string); v != "" {
-					return v == userID
-				}
-				return false
-			}
-			return false
-		}
-		return payload["user_id"] == userID
-	}
-	return true
-}
-
-func eventRevokesWorkspaceAccess(event store.Event, userID string) bool {
-	if event.Type != "bot.deleted" && event.Type != "bot.membership_removed" {
-		return false
-	}
-	switch payload := event.Payload.(type) {
-	case map[string]string:
-		return payload["bot_user_id"] == userID
-	case map[string]any:
-		botUserID, _ := payload["bot_user_id"].(string)
-		return botUserID == userID
-	default:
-		return false
-	}
-}
-
-func filterEventsForUser(events []store.Event, userID string) []store.Event {
-	filtered := events[:0]
-	for _, event := range events {
-		if shouldDeliverEvent(event, userID) {
-			filtered = append(filtered, event)
-		}
-	}
-	return filtered
-}
-
-func (s *Server) shouldDeliverEventToActor(ctx context.Context, event store.Event, userID string) bool {
-	deliver, err := s.shouldDeliverEventToActorResult(ctx, event, userID)
-	return err == nil && deliver
-}
-
-func (s *Server) shouldDeliverEventToActorResult(ctx context.Context, event store.Event, userID string) (bool, error) {
-	if !shouldDeliverEvent(event, userID) {
-		return false, nil
-	}
-	var err error
-	if conversationID := directConversationIDFromEvent(event); conversationID != "" {
-		_, err = s.store.GetDirectConversation(ctx, conversationID, userID)
-	} else if event.ChannelID == "" {
-		_, err = s.store.GetWorkspace(ctx, event.WorkspaceID, userID)
-	} else {
-		_, err = s.store.GetChannel(ctx, event.ChannelID, userID)
-	}
-	if err == nil {
-		return true, nil
-	}
-	if errors.Is(err, sql.ErrNoRows) {
-		return false, nil
-	}
-	return false, err
-}
-
-func directConversationIDFromEvent(event store.Event) string {
-	switch payload := event.Payload.(type) {
-	case map[string]string:
-		return payload["direct_conversation_id"]
-	case map[string]any:
-		conversationID, _ := payload["direct_conversation_id"].(string)
-		return conversationID
-	default:
-		return ""
-	}
-}
-
-var errSessionLookupUnavailable = errors.New("session verification unavailable; retry later")
-
-// Only missing or expired sessions invalidate authentication. Backend failures
-// must remain retryable across HTTP and already-open realtime connections.
-func (s *Server) sessionUser(ctx context.Context, token string) (store.User, error) {
-	user, err := s.store.GetSessionUser(ctx, token)
-	if err != nil && !errors.Is(err, sql.ErrNoRows) && !errors.Is(err, store.ErrSessionExpired) {
-		return store.User{}, fmt.Errorf("%w: %w", errSessionLookupUnavailable, err)
-	}
-	return user, err
-}
-
-func (s *Server) currentActor(r *http.Request) (actor, error) {
-	if auth := r.Header.Get("Authorization"); strings.HasPrefix(auth, "Bearer ") {
-		token := strings.TrimSpace(strings.TrimPrefix(auth, "Bearer "))
-		if botAuth, err := s.store.GetBotTokenAuth(r.Context(), token); err == nil {
-			// Record ingress use once; realtime revalidation stays read-only.
-			_ = s.store.RecordBotTokenUse(r.Context(), botAuth.TokenID)
-			return actor{
-				user:        botAuth.User,
-				botTokenID:  botAuth.TokenID,
-				workspaceID: botAuth.WorkspaceID,
-				scopes:      botAuth.Scopes,
-			}, nil
-		}
-		user, err := s.sessionUser(r.Context(), token)
-		if err != nil {
-			return actor{}, err
-		}
-		return actor{user: user, sessionToken: token}, nil
-	}
-	cookie, err := requestCookie(r, s.cookies.Session)
-	if errors.Is(err, errAmbiguousCookie) {
-		return actor{}, err
-	}
-	if err == nil && cookie.Value != "" {
-		user, err := s.sessionUser(r.Context(), cookie.Value)
-		if err == nil {
-			return actor{user: user, sessionToken: cookie.Value}, nil
-		}
-		if errors.Is(err, errSessionLookupUnavailable) || s.access == nil || r.Header.Get(accessAssertionHeader) == "" {
-			return actor{}, err
-		}
-	}
-	if s.access != nil {
-		if assertion := r.Header.Get(accessAssertionHeader); assertion != "" {
-			if act, err := s.accessActor(r, assertion); err == nil {
-				return act, nil
-			}
-		}
-	}
-	if s.disableDevAuth {
-		return actor{}, errors.New("authentication required")
-	}
-	if !isLocalDevRequest(r) {
-		return actor{}, errors.New("authentication required")
-	}
-	if id := r.Header.Get("X-ClickClack-User"); id != "" {
-		user, err := s.store.GetUser(r.Context(), id)
-		if err == nil && user.DeletedAt != nil {
-			return actor{}, errors.New("authentication required")
-		}
-		return actor{user: user}, err
-	}
-	user, err := s.store.FirstUser(r.Context())
-	return actor{user: user}, err
-}
-
-func (a actor) requireScope(scope string) error {
-	if a.botTokenID == "" {
-		return nil
-	}
-	for _, candidate := range a.scopes {
-		if candidate == scope {
-			return nil
-		}
-	}
-	return errors.New("bot token is missing scope " + scope)
-}
-
-func (a actor) requireWorkspace(workspaceID string) error {
-	if a.botTokenID == "" {
-		return nil
-	}
-	if a.workspaceID == workspaceID {
-		return nil
-	}
-	return errors.New("bot token cannot access this workspace")
-}
-
-func (s *Server) serveSPA(w http.ResponseWriter, r *http.Request) {
-	if r.URL.Path == "/api" || strings.HasPrefix(r.URL.Path, "/api/") {
-		writeError(w, http.StatusNotFound, errors.New("route not found"))
-		return
-	}
-
-	dist, err := fs.Sub(webassets.Dist, "dist")
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err)
-		return
-	}
-	if r.URL.Path != "/" {
-		if file, err := dist.Open(strings.TrimPrefix(r.URL.Path, "/")); err == nil {
-			_ = file.Close()
-			http.FileServer(http.FS(dist)).ServeHTTP(w, r)
-			return
-		}
-	}
-	if isMissingBrowserAssetPath(r.URL.Path) {
-		http.NotFound(w, r)
-		return
-	}
-	fallback := "index.html"
-	if r.URL.Path != "/" {
-		if _, err := fs.Stat(dist, "200.html"); err == nil {
-			fallback = "200.html"
-		}
-	}
-	index, err := fs.ReadFile(dist, fallback)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	w.Header().Set("Cache-Control", "no-store")
-	if strings.HasPrefix(r.URL.Path, "/embed/") {
-		// frame-ancestors is deliberately independent of cookie SameSite policy:
-		// allowing a cross-site ancestor never loosens cookies, so such embeds can
-		// render signed-out. Documented in docs/features/embedding.md.
-		ancestors := append([]string{"'self'"}, s.embedFrameAncestors...)
-		w.Header().Set("Content-Security-Policy", "frame-ancestors "+strings.Join(ancestors, " "))
-	}
-	index = s.injectRuntimeConfig(index)
-	_, _ = w.Write(index)
-}
-
-// enabledAuthMethods tells the frontend which sign-in surfaces to render. It
-// is always a non-nil slice so the SPA can distinguish "no method configured"
-// from an older server that omitted the field.
-func (s *Server) enabledAuthMethods() []string {
-	methods := []string{}
-	if s.githubOAuth.ClientID != "" && s.githubOAuth.ClientSecret != "" {
-		methods = append(methods, "github")
-	}
-	if s.passwordAuthEnabled {
-		methods = append(methods, "password")
-	}
-	return methods
-}
-
-func isMissingBrowserAssetPath(urlPath string) bool {
-	if strings.HasPrefix(urlPath, "/_app/") || strings.HasPrefix(urlPath, "/assets/") {
-		return true
-	}
-	switch strings.ToLower(path.Ext(urlPath)) {
-	case ".avif", ".css", ".gif", ".ico", ".jpeg", ".jpg", ".js", ".json",
-		".map", ".mjs", ".otf", ".png", ".svg", ".ttf", ".wasm", ".webmanifest",
-		".webp", ".woff", ".woff2":
-		return true
-	default:
-		return false
-	}
-}
-
-func (s *Server) injectRuntimeConfig(index []byte) []byte {
-	config, err := json.Marshal(struct {
-		APIBaseURL      string   `json:"apiBaseUrl"`
-		FrontendBaseURL string   `json:"frontendBaseUrl"`
-		AuthMethods     []string `json:"authMethods"`
-	}{
-		APIBaseURL:      s.publicAPIURL,
-		FrontendBaseURL: s.frontendURL,
-		AuthMethods:     s.enabledAuthMethods(),
-	})
-	if err != nil {
-		return index
-	}
-	script := append([]byte(`<script>window.__CLICKCLACK_CONFIG__=`), config...)
-	script = append(script, []byte(`;</script></head>`)...)
-	return bytes.Replace(index, []byte("</head>"), script, 1)
-}
-
-func writeWS(ctx context.Context, conn *websocket.Conn, event store.Event) error {
-	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-	body, err := json.Marshal(event)
-	if err != nil {
-		return err
-	}
-	return conn.Write(ctx, websocket.MessageText, body)
-}
-
-func readJSON(w http.ResponseWriter, r *http.Request, out any) error {
-	defer r.Body.Close()
-	r.Body = http.MaxBytesReader(w, r.Body, maxJSONBodyBytes)
-	decoder := json.NewDecoder(r.Body)
-	err := decoder.Decode(out)
-	if err != nil {
-		return err
-	}
-	var extra any
-	if err := decoder.Decode(&extra); err == nil {
-		if _, drainErr := io.Copy(io.Discard, r.Body); drainErr != nil {
-			return drainErr
-		}
-		return errors.New("json request body must contain a single JSON value")
-	} else if !errors.Is(err, io.EOF) {
-		return err
-	}
-	return nil
-}
-
-func writeResult(w http.ResponseWriter, body any, err error) {
-	writeResultStatus(w, http.StatusOK, body, err)
-}
-
-func writeResultStatus(w http.ResponseWriter, status int, body any, err error) {
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	writeJSON(w, status, body)
-}
-
-func writeMessageCreateResult(w http.ResponseWriter, message store.Message, event store.Event, err error) {
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	body := map[string]any{"message": message}
-	status := http.StatusOK
-	if event.ID != "" {
-		body["event"] = event
-		status = http.StatusCreated
-	}
-	writeJSON(w, status, body)
-}
-
-func writeThreadReplyCreateResult(w http.ResponseWriter, message store.Message, state store.ThreadState, events []store.Event, err error) {
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	status := http.StatusOK
-	if len(events) > 0 {
-		status = http.StatusCreated
-	}
-	writeJSON(w, status, map[string]any{"message": message, "thread_state": state, "events": events})
-}
-
-func (s *Server) writeReactionMutationResult(w http.ResponseWriter, r *http.Request, changedStatus int, userID, messageID string, event store.Event, err error) {
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	message, err := s.store.GetMessage(r.Context(), messageID, userID)
-	if err != nil {
-		writeStoreError(w, err)
-		return
-	}
-	status := http.StatusOK
-	if event.ID != "" {
-		status = changedStatus
-	}
-	writeJSON(w, status, map[string]any{"event": event, "reactions": message.Reactions})
-}
-
-func writeJSON(w http.ResponseWriter, status int, body any) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	_ = json.NewEncoder(w).Encode(body)
-}
-
-func writeError(w http.ResponseWriter, status int, err error) {
-	if errors.Is(err, errSessionLookupUnavailable) {
-		status = http.StatusServiceUnavailable
-		err = errSessionLookupUnavailable
-	}
-	var maxBytesErr *http.MaxBytesError
-	if errors.As(err, &maxBytesErr) {
-		status = http.StatusRequestEntityTooLarge
-	}
-	writeJSON(w, status, map[string]any{"error": err.Error()})
-}
-
-func writeStoreError(w http.ResponseWriter, err error) {
-	switch {
-	case errors.Is(err, store.ErrPostRateLimited):
-		writeError(w, http.StatusTooManyRequests, err)
-	case errors.Is(err, store.ErrUploadQuotaExceeded):
-		writeError(w, http.StatusRequestEntityTooLarge, err)
-	case errors.Is(err, store.ErrUploadNonceConflict):
-		writeError(w, http.StatusConflict, err)
-	case errors.Is(err, store.ErrSetupNonceConflict):
-		writeError(w, http.StatusConflict, err)
-	case errors.Is(err, store.ErrAlreadyPinned):
-		writeError(w, http.StatusConflict, err)
-	case errors.Is(err, store.ErrPinnedMessageLimit):
-		writeError(w, http.StatusConflict, err)
-	case errors.Is(err, store.ErrSetupCodeInvalid):
-		writeError(w, http.StatusNotFound, err)
-	case errors.Is(err, store.ErrModerationRestricted):
-		writeError(w, http.StatusForbidden, err)
-	case errors.Is(err, store.ErrNotWorkspaceManager):
-		writeError(w, http.StatusForbidden, err)
-	case errors.Is(err, store.ErrWorkspaceOwnerRequired):
-		writeError(w, http.StatusForbidden, err)
-	case errors.Is(err, store.ErrBotOwnerRequired):
-		writeError(w, http.StatusForbidden, err)
-	case errors.Is(err, store.ErrBotOwnerMembershipRequired):
-		writeError(w, http.StatusForbidden, err)
-	case errors.Is(err, store.ErrBotOwnerCreateRequired):
-		writeError(w, http.StatusForbidden, err)
-	case errors.Is(err, store.ErrMessageNotWritable):
-		writeError(w, http.StatusForbidden, err)
-	case errors.Is(err, store.ErrDirectConversationNoActivePeer):
-		writeError(w, http.StatusConflict, err)
-	default:
-		writeError(w, http.StatusBadRequest, err)
-	}
-}
-
-// optionalString returns a non-empty trimmed pointer or nil. Useful for JSON
-// fields that should map to a nullable Go pointer when absent or blank.
-func optionalString(value string) *string {
-	trimmed := strings.TrimSpace(value)
-	if trimmed == "" {
-		return nil
-	}
-	return &trimmed
-}
-
-func queryInt(r *http.Request, key string, fallback int) int {
-	value, err := strconv.ParseInt(r.URL.Query().Get(key), 10, 32)
-	if err != nil {
-		return fallback
-	}
-	return int(value)
-}
-
-func parseMessagePageRequest(r *http.Request) (store.MessagePageRequest, error) {
-	values := r.URL.Query()
-	req := store.MessagePageRequest{Limit: queryInt(r, "limit", 100)}
-	cursorCount := 0
-	for _, cursor := range []struct {
-		key string
-		set func(int64)
-	}{
-		{"before_seq", func(v int64) { req.BeforeSeq = &v }},
-		{"after_seq", func(v int64) { req.AfterSeq = &v }},
-		{"around_seq", func(v int64) { req.AroundSeq = &v }},
-	} {
-		raw, ok := values[cursor.key]
-		if !ok {
-			continue
-		}
-		cursorCount++
-		if len(raw) == 0 || strings.TrimSpace(raw[0]) == "" {
-			return req, fmt.Errorf("%w: %s is required", store.ErrInvalidMessagePage, cursor.key)
-		}
-		value, err := strconv.ParseInt(raw[0], 10, 64)
-		if err != nil || value < 0 {
-			return req, fmt.Errorf("%w: %s must be a non-negative integer", store.ErrInvalidMessagePage, cursor.key)
-		}
-		cursor.set(value)
-	}
-	if cursorCount > 1 {
-		return req, fmt.Errorf("%w: before_seq, after_seq, and around_seq are mutually exclusive", store.ErrInvalidMessagePage)
-	}
-	if mode := values.Get("mode"); mode != "" {
-		if mode != "latest" {
-			return req, fmt.Errorf("%w: unsupported message page mode %q", store.ErrInvalidMessagePage, mode)
-		}
-		if cursorCount > 0 {
-			return req, fmt.Errorf("%w: mode and cursor params are mutually exclusive", store.ErrInvalidMessagePage)
-		}
-	}
-	return req, nil
-}
-
-func writeMessagePage(w http.ResponseWriter, page store.MessagePage, err error) {
-	writeResult(w, map[string]any{
-		"messages":   page.Messages,
-		"oldest_seq": page.OldestSeq,
-		"newest_seq": page.NewestSeq,
-		"has_older":  page.HasOlder,
-		"has_newer":  page.HasNewer,
-	}, err)
 }
 
 func ListenAndServe(ctx context.Context, addr string, handler http.Handler) error {
